+++ src/main/java/com/mraof/minestuck/block/EmergingCruxiteDowelBlock.java
@@ -1,23 +1,23 @@
 package com.mraof.minestuck.block;
 
 import com.mraof.minestuck.blockentity.ItemStackBlockEntity;
+import com.mraof.minestuck.item.MSItems;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
-import net.minecraft.util.StringRepresentable;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.EntityBlock;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.StateDefinition;
-import net.minecraft.world.level.block.state.properties.EnumProperty;
 import net.minecraft.world.level.material.PushReaction;
 import net.minecraft.world.level.storage.loot.LootContext;
 import net.minecraft.world.level.storage.loot.parameters.LootContextParams;
@@ -32,23 +32,20 @@ import java.util.List;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-public class CruxiteDowelBlock extends Block implements EntityBlock
+public class EmergingCruxiteDowelBlock extends Block implements EntityBlock
 {
-	public static final VoxelShape DOWEL_SHAPE = Block.box(5, 0, 5, 11, 8, 11);
+	public static final VoxelShape CRUXTRUDER_SHAPE = Block.box(5, 0, 5, 11, 5, 11);
 	
-	public static final EnumProperty<Type> DOWEL_TYPE = MSProperties.DOWEL_BLOCK;
-	
-	public CruxiteDowelBlock(Properties properties)
+	public EmergingCruxiteDowelBlock(Properties properties)
 	{
 		super(properties);
-		registerDefaultState(this.stateDefinition.any().setValue(DOWEL_TYPE, Type.DOWEL));
 	}
 	
 	@Override
 	@SuppressWarnings("deprecation")
 	public VoxelShape getShape(BlockState state, BlockGetter level, BlockPos pos, CollisionContext context)
 	{
-		return DOWEL_SHAPE;
+		return CRUXTRUDER_SHAPE;
 	}
 	
 	@Nullable
@@ -56,7 +53,7 @@ public class CruxiteDowelBlock extends Block implements EntityBlock
 	public BlockEntity newBlockEntity(BlockPos pos, BlockState state)
 	{
 		ItemStackBlockEntity be = new ItemStackBlockEntity(pos, state);
-		be.setStack(new ItemStack(this));
+		be.setStack(new ItemStack(MSItems.CRUXITE_DOWEL.get()));
 		return be;
 	}
 	
@@ -76,15 +73,24 @@ public class CruxiteDowelBlock extends Block implements EntityBlock
 	public InteractionResult use(BlockState state, Level level, BlockPos pos, Player player, InteractionHand hand, BlockHitResult hit)
 	{
 		if(!level.isClientSide)
-			dropDowel(level, pos);
+			CruxiteDowelBlock.dropDowel(level, pos);
 		return  InteractionResult.SUCCESS;
 	}
 	
-	@Nullable
+	@SuppressWarnings("deprecation")
 	@Override
-	public BlockState getStateForPlacement(BlockPlaceContext context)
+	public BlockState updateShape(BlockState state, Direction direction, BlockState neighborState, LevelAccessor level, BlockPos currentPos, BlockPos neighborPos)
 	{
-		return context.getClickedFace() == Direction.UP ? defaultBlockState() : null;
+		return direction == Direction.DOWN && !this.canSurvive(state, level, currentPos)
+				? Blocks.AIR.defaultBlockState()
+				: super.updateShape(state, direction, neighborState, level, currentPos, neighborPos);
+	}
+	
+	@SuppressWarnings("deprecation")
+	@Override
+	public boolean canSurvive(BlockState state, LevelReader level, BlockPos pos)
+	{
+		return level.getBlockState(pos.below()).is(MSBlocks.CRUXTRUDER.TUBE.get());
 	}
 	
 	@Override
@@ -96,23 +102,7 @@ public class CruxiteDowelBlock extends Block implements EntityBlock
 			if(!dowel.isEmpty())
 				return dowel.copy();
 		}
-		return super.getCloneItemStack(state, target, level, pos, player);
-	}
-	
-	public static void dropDowel(Level level, BlockPos pos)
-	{
-		if (level.getBlockEntity(pos) instanceof ItemStackBlockEntity stackEntity)
-		{
-			ItemStack stack = stackEntity.getStack();
-			popResource(level, pos, stack);
-		}
-		level.removeBlock(pos, false);
-	}
-	
-	@Override
-	protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder)
-	{
-		builder.add(DOWEL_TYPE);
+		return new ItemStack(MSItems.CRUXITE_DOWEL.get());
 	}
 	
 	@Override
@@ -120,18 +110,5 @@ public class CruxiteDowelBlock extends Block implements EntityBlock
 	public PushReaction getPistonPushReaction(BlockState state)
 	{
 		return PushReaction.DESTROY;
-	}
-	
-	public enum Type implements StringRepresentable
-	{
-		DOWEL,
-		TOTEM;
-		
-		
-		@Override
-		public String getSerializedName()
-		{
-			return this.name().toLowerCase();
-		}
 	}
 }